implemented databse

+++ app/src/main/java/com/chiaching/todolistfragments/fragments/ui/HomeFragment.kt
@@ -1,7 +1,5 @@
 package com.chiaching.todolistfragments.fragments.ui
 
-import android.content.Intent
-import android.os.Build
 import android.os.Bundle
 import android.util.Log
 import androidx.fragment.app.Fragment
@@ -9,23 +7,13 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
-import androidx.activity.result.ActivityResultLauncher
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.annotation.RequiresApi
-import androidx.appcompat.app.AppCompatActivity
-import androidx.fragment.app.setFragmentResult
 import androidx.fragment.app.setFragmentResultListener
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.NavHostFragment
-import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.chiaching.todolistfragments.MyApplication
 import com.chiaching.todolistfragments.adapters.TaskAdapter
 import com.chiaching.todolistfragments.databinding.FragmentHomeBinding
-import com.chiaching.todolistfragments.model.Task
 import com.chiaching.todolistfragments.viewModels.HomeViewModel
 
 class HomeFragment : Fragment() {
@@ -84,7 +72,7 @@ class HomeFragment : Fragment() {
     fun setupAdapter(){
         val layoutManager = LinearLayoutManager(requireContext())
         adapter = TaskAdapter(emptyList()){
-            val action = HomeFragmentDirections.actionHomeFragmentToDetailsFragment(it.id!!)
+            val action = HomeFragmentDirections.actionHomeFragmentToDetailsFragment(it.id!!.toLong())
             NavHostFragment.findNavController(this).navigate(action)
         }
         binding.rvTask.adapter = adapter
